+++ changelog.md
@@ -6,3 +6,8 @@
 ## v1.3.0
 * adds `spacetime.whereIts()` method
 * actually implement 😓 season by hemisphere
+
+## v1.4.0
+* adds `spacetime.i18n()` method
+* adds `spacetime.nearest()` method
+* support for `"quarterHour"` units - '4:15, 4:30, 4:45, 5:00' etc